+++ packages/wepin_flutter_widget_sdk/example/lib/widgets/email_login_dialog.dart
@@ -1,9 +1,14 @@
 import 'package:flutter/material.dart';
 
 class EmailLoginDialog extends StatefulWidget {
-  final Function(String email, String password) onLogin;
+  final bool requirePassword; // Control whether to require password or not
+  final Function onLogin; // Function that can take different parameters based on requirePassword
 
-  const EmailLoginDialog({super.key, required this.onLogin});
+  const EmailLoginDialog({
+    super.key,
+    required this.onLogin,
+    this.requirePassword = true, // Default is to require a password
+  });
 
   @override
   EmailLoginDialogState createState() => EmailLoginDialogState();
@@ -24,17 +29,22 @@ class EmailLoginDialogState extends State<EmailLoginDialog> {
             onChanged: (value) => email = value,
             decoration: const InputDecoration(labelText: 'Email'),
           ),
-          TextField(
-            onChanged: (value) => password = value,
-            decoration: const InputDecoration(labelText: 'Password'),
-            obscureText: true,
-          ),
+          if (widget.requirePassword)
+            TextField(
+              onChanged: (value) => password = value,
+              decoration: const InputDecoration(labelText: 'Password'),
+              obscureText: true,
+            ),
         ],
       ),
       actions: [
         TextButton(
           onPressed: () {
-            widget.onLogin(email, password);
+            if (widget.requirePassword) {
+              widget.onLogin(email, password); // Call with email and password
+            } else {
+              widget.onLogin(email: email); // Call with email only
+            }
             Navigator.of(context).pop();
           },
           child: const Text('Login'),